% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{GetMaxRho}
\alias{GetMaxRho}
\title{GetMaxRho}
\usage{
GetMaxRho(X, Y, L, Omega0)
}
\arguments{
\item{X}{matrix of SNPs}

\item{Y}{gene expression matrix}

\item{L}{matrix of hidden factors}

\item{Omega0}{Boolean matrix of observed entries}
}
\description{
\code{GetMaxRho} is a function used to determine the maximum value as a candidate for rho.  See the parmeter tuning section of Yang et al. (2013)
Note:  This function is adapted from the LORS MATLAB implementation
}
\examples{

##Example

## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E
Omega0 <- !(is.na(Y))

## Usage
GetMaxRho(X, Y, L, Omega0)
}
