% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{HC_Screening}
\alias{HC_Screening}
\title{HC_Screening
\code{HC_Screening} is a function to apply the HC-Screening screening method of Rhyne et al. (2018) (In progress)
Note:  HC-Screening ranks SNPs by their higher criticism statistics and selects the top n, where n is the number of samples}
\usage{
HC_Screening(Y, X)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}
}
\value{
selectedSNPs The SNPs selected by HC-Screening
}
\description{
HC_Screening
\code{HC_Screening} is a function to apply the HC-Screening screening method of Rhyne et al. (2018) (In progress)
Note:  HC-Screening ranks SNPs by their higher criticism statistics and selects the top n, where n is the number of samples
}
\examples{

## Example

## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(q*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage
HC_Screening(Y, X)


}
