% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{InitialEst}
\alias{InitialEst}
\title{InitialEst
\code{InitialEst} is a function to build an initial estimate for B}
\usage{
InitialEst(Y, X, lambda = NULL)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{lambda}{tuning parameter}
}
\description{
InitialEst
\code{InitialEst} is a function to build an initial estimate for B
}
\examples{

}
