% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{linspace}
\alias{linspace}
\title{linspace}
\usage{
linspace(x1, x2, n = 100)
}
\arguments{
\item{x1}{a starting point}

\item{x2}{an ending point}

\item{n}{length of sequence}
}
\description{
\code{linspace} is a function to space a sequence linearly from x1 to x2
}
\examples{
##Example
linspace(100,10,5)
}
