% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{LORS0}
\alias{LORS0}
\title{LORS0}
\usage{
LORS0(Y, X, rho, lambda, maxiter = 1000, eps = 2.2204e-16, tol = 1e-04,
  verbose = FALSE)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{rho}{parameter for enforcing sparsity of coefficient matrix}

\item{lambda}{parameter for enforcing low-rank structure of hidden factor matrix}

\item{maxiter}{maximum number of iterations}

\item{eps}{constant used when checking the convergence.  Ensures no division by 0.}

\item{tol}{tolerance level for convergence}

\item{verbose}{chooses whether details should be printed to console.  Default is FALSE.}
}
\value{
\item{B}{The estimated coefficients} \item{mu}{The estimated intercept} \item{L}{The estimated matrix of hidden factors} \item{f_val_vec}{The objective function values} \item{res_vec}{The relative change in objective function values}
}
\description{
\code{LORS0} is a function for solving the LORS optimization problem through the method described in Can Yang et al. (2013).
This function is adapted from the authors MATLAB implementation
}
\examples{

##Example

#' ## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(q*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

rho <- runif(1,3,5)
lambda <- runif(1,3,5)
LORS0(Y, X, rho, lambda)
}
