% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{LORS2}
\alias{LORS2}
\title{LORS2}
\usage{
LORS2(Y, X, L, Omega1, Omega2, B, rho, lambda, tol, maxIter = 1000)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{L}{matrix of hidden factors}

\item{Omega1}{Boolean matrix for training data}

\item{Omega2}{Boolean matrix for validation data}

\item{rho}{parameter for enforcing sparsity of coefficient matrix}

\item{lambda}{parameter for enforcing low-rank structure of hidden factor matrix}

\item{tol}{tolerance level for convergence}
}
\description{
\code{LORS2} is a function used in parameter tuning in LORS. See the parameter tuning section described in Can Yang et al. (2013).
This function is adapted from the authors MATLAB implementation
}
\examples{

##Example

#' ## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

Omega0 <- Omega0 = !(is.na(Y))
mask <- matrix(runif(nrow(Y)*ncol(Y)) > 0.5, nrow = nrow(Y), ncol = ncol(Y))
Omega1 <- Omega0 & mask
Omega2 <- Omega0 & !mask
rho <- runif(1,3,5)
lambda <- runif(1,3,5)
tol <- 1e-4

## Usage
LORS2(Y, X, L, Omega1, Omega2, B, rho, lambda, tol)
}
