% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{LORS_Screen_Parallel}
\alias{LORS_Screen_Parallel}
\title{LORS_Screen_Parallel
\code{LORS_Screen_Parallel} is a function used to run LORS-Screening on a subset of the columns of X.  Can be used to perform LORS-Screening in parallel on a cluster.}
\usage{
LORS_Screen_Parallel(Y, X, chunk)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{chunk}{a group of columns to run the screening on.  Done in batches of 1000.}
}
\value{
\item{myB}{matrix of coefficients from LORS-Screening} \item{lambda}{tuning parameter used in LORS-Screening}
}
\description{
LORS_Screen_Parallel
\code{LORS_Screen_Parallel} is a function used to run LORS-Screening on a subset of the columns of X.  Can be used to perform LORS-Screening in parallel on a cluster.
}
\examples{

##Example

## Generate some data
n <- 20
p <- 50
q <- 30
k <- 4
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage
LORS_Screen_Parallel(Y, X, chunk = 1)
}
