% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{LORSscreen}
\alias{LORSscreen}
\title{LORSscreen
\code{LORSscreen} is a function to solve the LORS-Screening optimization problem in Yang et al. (2013)}
\usage{
LORSscreen(Y, X, lambda, tol)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{lambda}{tuning parameter}

\item{tol}{a tolerance level}
}
\description{
LORSscreen
\code{LORSscreen} is a function to solve the LORS-Screening optimization problem in Yang et al. (2013)
}
\examples{


}
