% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{mysub}
\alias{mysub}
\title{mysub}
\usage{
mysub(A, Omega0, j, type)
}
\arguments{
\item{A}{matrix}

\item{Omega0}{Boolean matrix of indices}

\item{j}{index for columns}

\item{type}{chooses the type of indices to collect. 1 corresponds to A(Omega0(:,j),:) in MATLAB and 2 corresponds to A(Omega0(:,j),j)}
}
\description{
\code{mysub} is an internal function used in parameter tuning to select indices of training/validation data
}
