% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{ParamTuneParallel}
\alias{ParamTuneParallel}
\title{ParamTuneParallel
\code{ParamTuneParallel} is a function used to run the parameter tuning of either FastLORS or LORS in parallel}
\usage{
ParamTuneParallel(Y, X, fold, seed = 123)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{fold}{fold in cross-validation}

\item{seed}{random seed used to create training and validation sets}
}
\value{
\item{myParams}{lambda and a sequence of rho values to use in parameter tuning} \item{Training}{Training Data} \item{Validation}{Validation Data}
}
\description{
ParamTuneParallel
\code{ParamTuneParallel} is a function used to run the parameter tuning of either FastLORS or LORS in parallel
}
\examples{

##Example

## Generate some data
n <- 20
p <- 50
q <- 30
k <- 4
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage
ParamTuneParallel(Y, X, fold = 1)
}
