% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{prox_1}
\alias{prox_1}
\title{prox_1}
\usage{
prox_1(b, tau)
}
\arguments{
\item{b}{a matrix or vector}

\item{tau}{the value to to apply soft thresholding with}
}
\description{
\code{prox_1} is the soft-thresholding function
}
