% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{rankHC}
\alias{rankHC}
\title{rankHC
\code{rankHC} is a function used to rank a Bhat matrix by higher criticism statistics}
\usage{
rankHC(Bhat_standardized)
}
\arguments{
\item{Bhat_standardized}{a standardized coefficient matrix}
}
\description{
rankHC
\code{rankHC} is a function used to rank a Bhat matrix by higher criticism statistics
}
