% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{Run_LORS}
\alias{Run_LORS}
\title{Run_LORS}
\usage{
Run_LORS(Y, X, method = "Fast-LORS", screening = "HC-Screening",
  tune_method = "Fast-LORS", seed = 123, maxiter = 10000,
  eps = 2.2204e-16, tol = 1e-04, cross_valid = TRUE)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{method}{chooses with modeling method to run}

\item{screening}{chooses whether LORS-Screening or HC-Screening should be performed.  Recommended if the number of SNPs is large.}

\item{maxiter}{maximum number of iterations}

\item{eps}{constant used when checking the convergence.  Ensures no division by 0.}

\item{tol}{tolerance level for convergence}
}
\description{
\code{Run_LORS} is a function used to run either Fast-LORS or LORS
}
\examples{

##Example

## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage
Run_LORS(Y, X, method = "Fast-LORS")

}
