% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{Run_LORS_Screening}
\alias{Run_LORS_Screening}
\title{Run_LORS_Screening
\code{Run_LORS_Screening} is a function to to apply the LORS-Screening Algorithm in Yang et al. (2013)}
\usage{
Run_LORS_Screening(Y, X, lambda = NULL)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{lambda}{tuning parameter}
}
\value{
selectedSNPs the SNPs selected by LORS-Screening
}
\description{
Run_LORS_Screening
\code{Run_LORS_Screening} is a function to to apply the LORS-Screening Algorithm in Yang et al. (2013)
}
\examples{

##Example

## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

selectedSNPs <- Run_LORS_Screening(Y, X)
}
