% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{S}
\alias{S}
\title{S
\code{S} is a function used internally in rankHC.  It calculates empirical cdf's.}
\usage{
S(t, my_matrix)
}
\arguments{
\item{t}{cutoff value of empirical cdf}

\item{my_matrix}{a coefficient matrix}
}
\value{
The empirical distribution function of the coefficients evaluated at t
}
\description{
S
\code{S} is a function used internally in rankHC.  It calculates empirical cdf's.
}
