% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{standardizeBhat}
\alias{standardizeBhat}
\title{standardizeBhat
\code{standardizeBhat} is a function used to standardize a coefficient matrix}
\usage{
standardizeBhat(Y, X, Bhat)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{Bhat}{a coefficient matrix}
}
\description{
standardizeBhat
\code{standardizeBhat} is a function used to standardize a coefficient matrix
}
