% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{SVT}
\alias{SVT}
\title{SVT}
\usage{
SVT(Y, lambda)
}
\arguments{
\item{Y}{gene expression matrix}

\item{lambda}{a tuning parameter}
}
\value{
L the singular value decomposition of Y, soft-thresholded with lambda.
}
\description{
\code{SVT} is a function to perform soft-thresholded singular value decomposition.  It is used to get an initial estimate for L.
Note:  This function is adapted from the LORS MATLAB implementation
}
\examples{

##Example
set.seed(123)

Y <-matrix(rnorm(50*100, 7,1), nrow = 50, ncol = 100)
lambda <- runif(1,3,5)
SVT(Y, lambda)
}
