% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{Fast_LORS_Tuning}
\alias{Fast_LORS_Tuning}
\title{Fast_LORS_Tuning}
\usage{
Fast_LORS_Tuning(Y, X, rho, lambda, Training, Validation, maxiter = 5000,
  eps = 2.2204e-16, tol = 1e-04, B = NULL, mu = NULL, L = NULL)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{rho}{parameter used to enforce sparsity of B}

\item{lambda}{parameter used to enforce low-rank structure of L}

\item{Training}{Boolean matrix for training data}

\item{Validation}{Boolean matrix for validation data}

\item{maxiter}{maximum number of iterates}

\item{eps}{a small constant to prevent dividing by zero when checking relative change in function values.}

\item{tol}{tolerance threshold for convergence}

\item{B}{an estimate for matrix of coefficients.  Default is NULL.}

\item{mu}{an estimate for the intercept.  Default is NULL.}

\item{L}{an estimate for the hidden factors  Default is NULL.}
}
\value{
\item{B}{matrix of coefficients} \item{L}{matrix of hidden factors} \item{mu}{vector of intercepts} \item{Err}{Residual sum of squares of validation data} \item{f_vals}{Objective function values} \item{res_vec}{Relative change in objective function values} \item{iter}{Number of iterations}
}
\description{
\code{Fast_LORS_Tuning} is a function used perform parameter tuning using FastLORS instead of LORS
}
