% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{InitialEst}
\alias{InitialEst}
\title{InitialEst
\code{InitialEst} is a function to build an initial estimate for B}
\usage{
InitialEst(Y, X, lambda = NULL)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{lambda}{tuning parameter}
}
\value{
B An initial estimate of the coefficient matrix
}
\description{
InitialEst
\code{InitialEst} is a function to build an initial estimate for B
}
\examples{

## Example
## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(q*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
  genes_influenced <- sort(sample(c(1:ncol(B)),5))
  B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

Init_est <- InitialEst(Y,X)
}
