% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{LORSscreen}
\alias{LORSscreen}
\title{LORSscreen
\code{LORSscreen} is a function to solve the LORS-Screening optimization problem in Yang et al. (2013)}
\usage{
LORSscreen(Y, X, lambda, tol)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{a SNP}

\item{lambda}{tuning parameter}

\item{tol}{a tolerance level}
}
\value{
B the estimated coefficients for the SNP

L the estimated hidden factors

mu the estimate for the intercept
}
\description{
LORSscreen
\code{LORSscreen} is a function to solve the LORS-Screening optimization problem in Yang et al. (2013)
}
\examples{

##Example

## Generate some data
n <- 50
p <- 200
q <- 100
k <- 10
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(n*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage to build initial estimate

Bhat_initial <- c()
for(SNP_col in 1:ncol(X)){
   X1 <- matrix(X[,SNP_col], ncol = 1)
   LS <- LORSscreen(Y, X1, lambda = 0.1, 0.01)
   B_row <- LS$B
   Bhat_initial <- rbind(Bhat_initial, B_row)
}
}
