% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{prox_1}
\alias{prox_1}
\title{prox_1}
\usage{
prox_1(b, tau)
}
\arguments{
\item{b}{a matrix or vector}

\item{tau}{the value to to apply soft thresholding with}
}
\value{
prox_b the soft-thresholding function applied to b with threshold tau
}
\description{
\code{prox_1} is the soft-thresholding function.  Let the entries of b be b_j.  The function subtracts tau from b_j for b_j > tau, sets b_j to 0 where abs(b_j) < tau, and adds tau where b_j < -tau.
}
