% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{rankHC}
\alias{rankHC}
\title{rankHC
\code{rankHC} is a function used to rank a Bhat matrix by higher criticism statistics}
\usage{
rankHC(Bhat_standardized)
}
\arguments{
\item{Bhat_standardized}{a standardized coefficient matrix}
}
\value{
\item{index}{The indices of the sorted Higher Criticism values} \item{HC_vec}{The higher criticism statistics for the standardized Bhat matrix}
}
\description{
rankHC
\code{rankHC} is a function used to rank a Bhat matrix by higher criticism statistics
}
