% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{Run_LORS}
\alias{Run_LORS}
\title{Run_LORS}
\usage{
Run_LORS(Y, X, method = "FastLORS", screening = "LORS-Screening",
  tune_method = "FastLORS", seed = 123, maxiter = 10000,
  eps = 2.2204e-16, tol = 1e-04, cross_valid = TRUE, omega_SOR = 1.999)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{method}{chooses with modeling method to run}

\item{screening}{Either "LORS-Screening", "HC-Screening", or "None".  The default method, LORS-Screening, is recommended if the number of SNPs is large. HC-Screening of Rhyne et al. (2018) is under development but is included here as an option.}

\item{tune_method}{chooses whether FastLORS should be used for parameter tuning or the original LORS procedure should be used.  Default is FastLORS}

\item{seed}{random seed to be used for setting training and validation set.  Default is 123.}

\item{maxiter}{maximum number of iterations}

\item{eps}{constant used when checking the convergence.  Ensures no division by 0.}

\item{tol}{tolerance level for convergence}

\item{cross_valid}{chooses whether cross-validation should be used in parameter tuning.  Default is TRUE.}

\item{omega_SOR}{the value of omega to use if applying successive over-relaxation with FastLORS.}
}
\value{
\item{LORS_Obj or Fast_LORS_Obj}{A list produced from LORS or FastLORS containing (1) B: estimate of the coefficient matrix (2) L:  estimate of the matrix of hidden factors (3) mu: estiamte of the vector of intercepts (4) f_val_vec: objective function values and (5) res_vec: relative change in objective function values} \item{selectedSNPs}{The SNPs selected by the screening method} \item{screening_time}{The time (in seconds) spent on screening step} \item{param_time}{The time (in seconds) spent on the parameter tuning step} \item{model_time}{The time (in seconds) spent on the joint modeling step} \item{total_time}{The time (in seconds) spent on the screening, parameter tuning, and joint modeling steps}  \item{rho}{The value of rho chosen through parameter tuning} \item{lambda}{The value of lambda chosen through parameter tuning}
}
\description{
\code{Run_LORS} is a function used to run either FastLORS or LORS
}
\examples{

##Example

## Generate some data
n <- 20
p <- 50
q <- 30
k <- 4
set.seed(123)
X <- matrix(rbinom(n*p,1,0.5),n,p)
L <- matrix(rnorm(n*k),n,k) \%*\% t(matrix(rnorm(q*k),q,k))
B <- matrix(0, ncol(X), ncol(L))
activeSNPs <- sort(sample(c(1:nrow(B)), 20))
for(i in 1:length(activeSNPs)){
genes_influenced <- sort(sample(c(1:ncol(B)),5))
B[activeSNPs[i], genes_influenced] <- 2
}
E <- matrix(rnorm(n*q),n,q)
Y <- X \%*\% B + L + E

## Usage
Run_LORS(Y, X, method = "FastLORS")
}
