% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{Run_LORS_Screening}
\alias{Run_LORS_Screening}
\title{Run_LORS_Screening
\code{Run_LORS_Screening} is a function to to apply the LORS-Screening Algorithm in Yang et al. (2013)}
\usage{
Run_LORS_Screening(Y, X, lambda = NULL)
}
\arguments{
\item{Y}{gene expression matrix}

\item{X}{matrix of SNPs}

\item{lambda}{tuning parameter}
}
\description{
Run_LORS_Screening
\code{Run_LORS_Screening} is a function to to apply the LORS-Screening Algorithm in Yang et al. (2013)
}
\examples{

}
