% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{softImpute}
\alias{softImpute}
\title{softImpute}
\usage{
softImpute(X, Z, Omega0, Omega1, Omega2, alpha0, maxRank)
}
\arguments{
\item{X}{a (possibly) incomplete matrix}

\item{Z}{the target matrix}

\item{Omega0}{Boolean matrix of observed entries}

\item{Omega1}{Boolean matrix of training entries}

\item{Omega2}{Boolean matrix of validation entries}

\item{alpha0}{initial tuning parameter}

\item{maxRank}{maximum rank of the solution}
}
\description{
\code{softImpute} is a function from Mazudmer et al. (2010). It solves the problem min || X - Z ||_Omega + alpha || Z ||_Nulear and is used in parameter tuning for LORS.
Note:  This function is adapted from the LORS MATLAB implementation
}
