% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS Functions.R
\name{survival}
\alias{survival}
\title{survival
\code{survival} is a function used internally in rankHC.  It calculates the survival function of the standard normal distribution}
\usage{
survival(t)
}
\arguments{
\item{t}{a cutoff value}
}
\description{
survival
\code{survival} is a function used internally in rankHC.  It calculates the survival function of the standard normal distribution
}
