% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastLORS_Functions.R
\name{svd_st}
\alias{svd_st}
\title{svd_st}
\usage{
svd_st(X, lambda)
}
\arguments{
\item{X}{a matrix}

\item{lambda}{the value to to apply soft thresholding with}
}
\value{
L the soft-thresholded singular value decomposition of X.
}
\description{
\code{svd_st} is a function for performing soft-thresholded singular value decomposition of a matrix X
}
